rename: common 컴포넌트 폴더명 @common으로 수정
