+++ src/components/@common/CroppedImage.tsx
